+++ pkg/cmd/pulumi/state.go
@@ -49,6 +49,7 @@ troubleshooting a stack or when performing specific edits that otherwise would r
 
 	cmd.AddCommand(newStateDeleteCommand())
 	cmd.AddCommand(newStateUnprotectCommand())
+	cmd.AddCommand(newStateProtectCommand())
 	return cmd
 }